refactor(spec-score): 완갑 장비 파워 테이블 분리
