+++ src/data/specScore/equipmentPowerTables.ts
@@ -1,3 +1,4 @@
+export * from './armletEquipmentPowerTables';
 export * from './equipmentPowerTableTypes';
 export * from './egirEquipmentPowerTables';
 export * from './serkaEquipmentPowerTables';
